Version 1.0.2

+++ flowapp/views/admin.py
@@ -1,13 +1,17 @@
-# flowapp/views/admin.py
+import csv
+from io import StringIO
 from datetime import datetime, timedelta
 import secrets
 
-from flask import Blueprint, render_template, redirect, flash, request, session, url_for
-from sqlalchemy.exc import IntegrityError
+from sqlalchemy import func, text
+from flask import Blueprint, render_template, redirect, flash, request, session, url_for, current_app
+import sqlalchemy
+from sqlalchemy.exc import IntegrityError, OperationalError
 
-from ..forms import ASPathForm, MachineApiKeyForm, UserForm, ActionForm, OrganizationForm, CommunityForm
+from ..forms import ASPathForm, BulkUserForm, MachineApiKeyForm, UserForm, ActionForm, OrganizationForm, CommunityForm
 from ..models import (
     ASPath,
+    ApiKey,
     MachineApiKey,
     User,
     Action,
@@ -18,6 +22,9 @@ from ..models import (
     Community,
     get_existing_community,
     Log,
+    Flowspec4,
+    Flowspec6,
+    RTBH,
 )
 from ..auth import auth_required, admin_required
 from flowapp import db
@@ -68,8 +75,6 @@ def add_machine_key():
     form = MachineApiKeyForm(request.form, key=generated)
 
     if request.method == "POST" and form.validate():
-        print("Form validated")
-        # import ipdb; ipdb.set_trace()
         model = MachineApiKey(
             machine=form.machine.data,
             key=form.key.data,
@@ -87,7 +92,7 @@ def add_machine_key():
     else:
         for field, errors in form.errors.items():
             for error in errors:
-                print("Error in the %s field - %s" % (getattr(form, field).label.text, error))
+                current_app.logger.debug("Error in the %s field - %s" % (getattr(form, field).label.text, error))
 
     return render_template("forms/machine_api_key.html", form=form, generated_key=generated)
 
@@ -100,7 +105,7 @@ def delete_machine_key(key_id):
     Delete api_key and machine
     :param key_id: integer
     """
-    model = db.session.query(MachineApiKey).get(key_id)
+    model = db.session.get(MachineApiKey, key_id)
     # delete from db
     db.session.delete(model)
     db.session.commit()
@@ -148,7 +153,7 @@ def user():
 @auth_required
 @admin_required
 def edit_user(user_id):
-    user = db.session.query(User).get(user_id)
+    user = db.session.get(User, user_id)
     form = UserForm(request.form, obj=user)
     form.role_ids.choices = [(g.id, g.name) for g in db.session.query(Role).order_by("name")]
     form.org_ids.choices = [(g.id, g.name) for g in db.session.query(Organization).order_by("name")]
@@ -163,7 +168,7 @@ def edit_user(user_id):
 
     return render_template(
         "forms/simple_form.html",
-        title="Editing {}".format(user.email),
+        title=f"Editing {user.email}",
         form=form,
         action_url=action_url,
     )
@@ -173,18 +178,27 @@ def edit_user(user_id):
 @auth_required
 @admin_required
 def delete_user(user_id):
-    user = db.session.query(User).get(user_id)
+    user = db.session.get(User, user_id)
+    if not user:
+        flash("User not found.", "alert-danger")
+        return redirect(url_for("admin.users"))
+
     username = user.email
     db.session.delete(user)
 
-    message = "User {} deleted".format(username)
+    message = f"User {username} deleted"
     alert_type = "alert-success"
+
     try:
         db.session.commit()
-    except IntegrityError as e:
-        message = "User {} owns some rules, can not be deleted!".format(username)
+    except IntegrityError:
+        db.session.rollback()  # Rollback on IntegrityError
+        message = f"User {username} owns some rules, cannot be deleted! Delete rules first."
         alert_type = "alert-danger"
-        print(e)
+    except OperationalError:
+        db.session.rollback()  # Rollback on OperationalError
+        message = f"User {username} owns some rules, cannot be deleted! Delete rules first."
+        alert_type = "alert-danger"
 
     flash(message, alert_type)
     return redirect(url_for("admin.users"))
@@ -198,12 +212,118 @@ def users():
     return render_template("pages/users.html", users=users)
 
 
+@admin.route("/bulk-import-users", methods=["GET"])
+@auth_required
+@admin_required
+def bulk_import_users():
+    form = BulkUserForm(request.form)
+    orgs = db.session.execute(db.select(Organization).order_by(Organization.name)).scalars()
+    return render_template("forms/bulk_user_form.html", form=form, orgs=orgs)
+
+
+@admin.route("/bulk-import-users", methods=["POST"])
+@auth_required
+@admin_required
+def bulk_import_users_save():
+    form = BulkUserForm(request.form)
+    roles = [role.id for role in db.session.query(Role).all()]
+    orgs = [org.id for org in db.session.query(Organization).all()]
+    uuids = [user.uuid for user in db.session.query(User).all()]
+    form.roles = roles
+    form.organizations = orgs
+    form.uuids = uuids
+
+    if request.method == "POST" and form.validate():
+        # Get CSV data from textarea
+        csv_data = form.users.data
+        # Parse CSV data
+        csv_reader = csv.DictReader(StringIO(csv_data), delimiter=",")
+        errored = False
+        for row in csv_reader:
+            try:
+                # Extract and prepare data
+                uuid = row["uuid-eppn"]
+                name = row["name"]
+                phone = row["telefon"]
+                email = row["email"]
+
+                # Convert role and organization fields to lists of integers
+                role_ids = [int(row["role"])]  # role_id should be a list
+                org_ids = [int(row["organizace"])]  # org_id should be a list
+                notice = row["poznamka"]
+
+                # Insert user
+                insert_user(
+                    uuid=uuid, role_ids=role_ids, org_ids=org_ids, name=name, phone=phone, email=email, comment=notice
+                )
+            except KeyError as e:
+                errored = True
+                # Handle missing fields or other errors in the CSV
+                flash(f"Missing field in CSV: {e}", "alert-danger")
+            except ValueError as e:
+                errored = True
+                # Handle conversion issues (like invalid int for role/org)
+                flash(f"Data conversion error: {e}", "alert-danger")
+            except sqlalchemy.exc.IntegrityError as e:
+                errored = True
+                db.session.rollback()
+                flash(f"SQL Integrity error: {e}", "alert-danger")
+
+        if not errored:
+            return redirect(url_for("admin.users"))
+
+    return render_template("forms/bulk_user_form.html", form=form)
+
+
 @admin.route("/organizations")
 @auth_required
 @admin_required
 def organizations():
-    orgs = db.session.query(Organization).all()
-    return render_template("pages/orgs.html", orgs=orgs)
+    # Query all organizations and eager load RTBH relationships
+    orgs = db.session.query(Organization).options(db.joinedload(Organization.rtbh)).all()
+
+    # Get RTBH counts with rstate_id=1 for all organizations in one query
+    rtbh_counts_query = (
+        db.session.query(RTBH.org_id, func.count(RTBH.id)).filter(RTBH.rstate_id == 1).group_by(RTBH.org_id).all()
+    )
+
+    flowspec4_count_query = (
+        db.session.query(Flowspec4.org_id, func.count(Flowspec4.id))
+        .filter(Flowspec4.rstate_id == 1)
+        .group_by(Flowspec4.org_id)
+        .all()
+    )
+
+    flowspec6_count_query = (
+        db.session.query(Flowspec6.org_id, func.count(Flowspec6.id))
+        .filter(Flowspec6.rstate_id == 1)
+        .group_by(Flowspec6.org_id)
+        .all()
+    )
+
+    flowspec4_all_count = db.session.query(Flowspec4).filter(Flowspec4.rstate_id == 1).count()
+    flowspec6_all_count = db.session.query(Flowspec6).filter(Flowspec6.rstate_id == 1).count()
+    rtbh_all_count = db.session.query(RTBH).filter(RTBH.rstate_id == 1).count()
+    flowspec_limit = current_app.config.get("FLOWSPEC_MAX_RULES", 9000)
+    rtbh_limit = current_app.config.get("RTBH_MAX_RULES", 100000)
+
+    # Convert query result to a dictionary {org_id: count}
+    rtbh_counts = {org_id: count for org_id, count in rtbh_counts_query}
+    flowspec4_counts = {org_id: count for org_id, count in flowspec4_count_query}
+    flowspec6_counts = {org_id: count for org_id, count in flowspec6_count_query}
+
+    return render_template(
+        "pages/orgs.html",
+        orgs=orgs,
+        rtbh_counts=rtbh_counts,
+        flowspec4_counts=flowspec4_counts,
+        flowspec6_counts=flowspec6_counts,
+        rtbh_all_count=rtbh_all_count,
+        flowspec4_all_count=flowspec4_all_count,
+        flowspec6_all_count=flowspec6_all_count,
+        flowspec_limit=flowspec_limit,
+        rtbh_limit=rtbh_limit,
+    )
 
 
 @admin.route("/organization", methods=["GET", "POST"])
@@ -216,7 +336,13 @@ def organization():
         # test if user is unique
         exist = db.session.query(Organization).filter_by(name=form.name.data).first()
         if not exist:
-            org = Organization(name=form.name.data, arange=form.arange.data)
+            org = Organization(
+                name=form.name.data,
+                arange=form.arange.data,
+                limit_flowspec4=form.limit_flowspec4.data,
+                limit_flowspec6=form.limit_flowspec6.data,
+                limit_rtbh=form.limit_rtbh.data,
+            )
             db.session.add(org)
             db.session.commit()
             flash("Organization saved")
@@ -226,7 +352,7 @@ def organization():
 
     action_url = url_for("admin.organization")
     return render_template(
-        "forms/simple_form.html",
+        "forms/org.html",
         title="Add new organization to Flowspec",
         form=form,
         action_url=action_url,
@@ -237,18 +363,18 @@ def organization():
 @auth_required
 @admin_required
 def edit_organization(org_id):
-    org = db.session.query(Organization).get(org_id)
+    org = db.session.get(Organization, org_id)
     form = OrganizationForm(request.form, obj=org)
 
     if request.method == "POST" and form.validate():
         form.populate_obj(org)
         db.session.commit()
-        flash("Organization updated")
+        flash("Organization updated", "alert-success")
         return redirect(url_for("admin.organizations"))
 
     action_url = url_for("admin.edit_organization", org_id=org.id)
     return render_template(
-        "forms/simple_form.html",
+        "forms/org.html",
         title="Editing {}".format(org.name),
         form=form,
         action_url=action_url,
@@ -259,7 +385,7 @@ def edit_organization(org_id):
 @auth_required
 @admin_required
 def delete_organization(org_id):
-    org = db.session.query(Organization).get(org_id)
+    org = db.session.get(Organization, org_id)
     aname = org.name
     db.session.delete(org)
     message = "Organization {} deleted".format(aname)
@@ -315,7 +441,7 @@ def as_path():
 @auth_required
 @admin_required
 def edit_as_path(path_id):
-    pth = db.session.query(ASPath).get(path_id)
+    pth = db.session.get(ASPath, path_id)
     form = ASPathForm(request.form, obj=pth)
 
     if request.method == "POST" and form.validate():
@@ -337,7 +463,7 @@ def edit_as_path(path_id):
 @auth_required
 @admin_required
 def delete_as_path(path_id):
-    pth = db.session.query(ASPath).get(path_id)
+    pth = db.session.get(ASPath, path_id)
     db.session.delete(pth)
     message = f"AS path {pth.prefix} : {pth.as_path} deleted"
     alert_type = "alert-success"
@@ -395,8 +521,7 @@ def action():
 @auth_required
 @admin_required
 def edit_action(action_id):
-    action = db.session.query(Action).get(action_id)
-    print(action.role_id)
+    action = db.session.get(Action, action_id)
     form = ActionForm(request.form, obj=action)
     if request.method == "POST" and form.validate():
         form.populate_obj(action)
@@ -417,7 +542,7 @@ def edit_action(action_id):
 @auth_required
 @admin_required
 def delete_action(action_id):
-    action = db.session.query(Action).get(action_id)
+    action = db.session.get(Action, action_id)
     aname = action.name
     db.session.delete(action)
 
@@ -480,8 +605,7 @@ def community():
 @auth_required
 @admin_required
 def edit_community(community_id):
-    community = db.session.query(Community).get(community_id)
-    print(community.role_id)
+    community = db.session.get(Community, community_id)
     form = CommunityForm(request.form, obj=community)
     if request.method == "POST" and form.validate():
         form.populate_obj(community)
@@ -502,7 +626,7 @@ def edit_community(community_id):
 @auth_required
 @admin_required
 def delete_community(community_id):
-    community = db.session.query(Community).get(community_id)
+    community = db.session.get(Community, community_id)
     aname = community.name
     db.session.delete(community)
     message = "Community {} deleted".format(aname)
@@ -515,3 +639,51 @@ def delete_community(community_id):
 
     flash(message, alert_type)
     return redirect(url_for("admin.communities"))
+
+
+@admin.route("/set-org-if-zero", methods=["GET"])
+@auth_required
+@admin_required
+def update_set_org():
+    # Define the raw SQL update statement
+    update_statement = update_statement = text(
+        """
+        UPDATE organization 
+        SET limit_flowspec4 = 0, limit_flowspec6 = 0, limit_rtbh = 0 
+        WHERE limit_flowspec4 IS NULL OR limit_flowspec6 IS NULL OR limit_rtbh IS NULL;
+    """
+    )
+    try:
+        # Execute the update query
+        db.session.execute(update_statement)
+        db.session.commit()
+    except Exception as e:
+        db.session.rollback()
+        flash(f"Error updating organizations: {e}", "alert-danger")
+
+    # Get all flowspec records where org_id is NULL (if this is needed)
+    models = [Flowspec4, Flowspec6, RTBH, ApiKey, MachineApiKey]
+    user_with_multiple_orgs = {}
+    for model in models:
+        data_records = model.query.filter(model.org_id == 0).all()
+        print(f"Found {len(data_records)} records with org_id NULL in {model.__name__}")
+        # Loop through each flowspec record and update org_id based on the user's organization
+        updated = 0
+        for row in data_records:
+            orgs = row.user.organization.all()
+            if len(orgs) == 1:
+                user_org = orgs[0]
+                if user_org:
+                    row.org_id = user_org.id
+                    updated += 1
+            else:
+                print(f"User {row.user.email} has multiple organizations")
+                user_with_multiple_orgs[row.user.email] = [org.name for org in orgs]
+        # Commit the changes
+        try:
+            db.session.commit()
+        except Exception as e:
+            db.session.rollback()
+            flash(f"Error updating {model.__name__}: {e}", "alert-danger")
+
+    return render_template("pages/user_list.html", users=user_with_multiple_orgs, updated=updated)
